[#S12P21E202-154][BE] Feat: 은행 서비스 및 페이 서비스 기능 구현

+++ BE/financial/src/main/java/com/ssafy/financial/dto/request/SimplePasswordVerifyRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class SimplePasswordVerifyRequest {
-    private Integer userId;         // 사용자 ID
+    private Long userId;         // 사용자 ID
     private String rawPassword;  // 사용자가 입력한 6자리 비밀번호
 }
